Path Update depending you role in the application
For Anonymous user Main menu link are the following : Fil d'Actu / Map /Inventaire /Participer

When you are connected Fil d'Actu is removed and changes by OBS' (link to the validate watching list view).
Fil d'Actu is still accessible clicking on the app's logo.
Connected as a Naturalist, Participer is removed (still accesible by a link in the footer) and a link Validation to watching to validate list view appairs.

On the circular menu the access to add observation path and to see the watching to validate page depends of your role and will locate you on the join us page if you haven't the good role level
Observation view.html.twig change -> the id obsCommentContent which has been removed in the previous version have been reintroduce (important for js script -> manage the display depending of the sreen)